moved the business logic from controller layer to service layer
