+++ service/src/main/java/org/cbioportal/genome_nexus/service/IsoformOverrideService.java
@@ -12,6 +12,7 @@ public interface IsoformOverrideService
 {
     IsoformOverride getIsoformOverride(String source, String id) throws IsoformOverrideNotFoundException;
     List<IsoformOverride> getIsoformOverrides(String source) throws IsoformOverrideNotFoundException;
+    List<IsoformOverride> getIsoformOverrides(String source, List<String> transcriptIds);
     List<String> getOverrideSources();
     Boolean hasData(String source);
 }
