+++ service/src/main/java/org/cbioportal/genome_nexus/service/VariantAnnotationService.java
@@ -36,6 +36,8 @@ import org.cbioportal.genome_nexus.model.VariantAnnotation;
 import org.cbioportal.genome_nexus.service.exception.VariantAnnotationNotFoundException;
 import org.cbioportal.genome_nexus.service.exception.VariantAnnotationWebServiceException;
 
+import java.util.List;
+
 /**
  * @author Benjamin Gross
  */
@@ -43,4 +45,8 @@ public interface VariantAnnotationService
 {
     VariantAnnotation getAnnotation(String variant)
         throws VariantAnnotationNotFoundException, VariantAnnotationWebServiceException;
+    List<VariantAnnotation> getAnnotations(List<String> variants);
+    VariantAnnotation getAnnotation(String variant, String isoformOverrideSource, List<String> fields)
+        throws VariantAnnotationWebServiceException, VariantAnnotationNotFoundException;
+    List<VariantAnnotation> getAnnotations(List<String> variants, String isoformOverrideSource, List<String> fields);
 }
