+++ web/src/main/java/org/cbioportal/genome_nexus/web/MutationAssessorController.java
@@ -4,22 +4,14 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.cbioportal.genome_nexus.model.MutationAssessor;
-import org.cbioportal.genome_nexus.model.MutationAssessorAnnotation;
-import org.cbioportal.genome_nexus.model.VariantAnnotation;
-import org.cbioportal.genome_nexus.service.AnnotationEnricher;
-import org.cbioportal.genome_nexus.service.EnrichmentService;
+import org.cbioportal.genome_nexus.service.MutationAssessorService;
 import org.cbioportal.genome_nexus.service.exception.VariantAnnotationNotFoundException;
 import org.cbioportal.genome_nexus.service.exception.VariantAnnotationWebServiceException;
-import org.cbioportal.genome_nexus.service.internal.MutationAssessorAnnotationEnricher;
-import org.cbioportal.genome_nexus.service.internal.MutationAssessorService;
-import org.cbioportal.genome_nexus.service.internal.VEPEnrichmentService;
 import org.cbioportal.genome_nexus.web.config.InternalApi;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 @InternalApi
 @RestController // shorthand for @Controller, @ResponseBody
@@ -28,14 +20,11 @@ import java.util.Map;
 @Api(tags = "mutation-assessor-controller", description = "Mutation Assessor Controller")
 public class MutationAssessorController
 {
-    private final VariantAnnotator variantAnnotator;
     private final MutationAssessorService mutationAssessorService;
 
     @Autowired
-    public MutationAssessorController(VariantAnnotator variantAnnotator,
-                                      MutationAssessorService mutationAssessorService)
+    public MutationAssessorController(MutationAssessorService mutationAssessorService)
     {
-        this.variantAnnotator = variantAnnotator;
         this.mutationAssessorService = mutationAssessorService;
     }
 
@@ -51,7 +40,7 @@ public class MutationAssessorController
         @PathVariable String variant)
         throws VariantAnnotationNotFoundException, VariantAnnotationWebServiceException
     {
-        return this.getMutationAssessorAnnotation(variant);
+        return this.mutationAssessorService.getMutationAssessorFromEnrichedVariantAnnotation(variant);
     }
 
     @ApiOperation(value = "Retrieves mutation assessor information for the provided list of variants",
@@ -65,89 +54,6 @@ public class MutationAssessorController
             allowMultiple = true)
         @RequestBody List<String> variants)
     {
-        return this.getMutationAssessorAnnotation(variants);
-    }
-
-    private MutationAssessor getMutationAssessorAnnotation(String variant)
-        throws VariantAnnotationNotFoundException, VariantAnnotationWebServiceException
-    {
-        VariantAnnotation variantAnnotation = this.getVariantAnnotation(variant);
-
-        return this.getMutationAssessorFromEnrichedVariantAnnotation(variantAnnotation);
-    }
-
-    private List<MutationAssessor> getMutationAssessorAnnotation(List<String> variants)
-    {
-        List<MutationAssessor> mutationAssessors = new ArrayList<>();
-
-        // uses enrichment to get mutation assessor object
-        List<String> fields = new ArrayList<>();
-        fields.add("mutation_assessor");
-
-        List<VariantAnnotation> variantAnnotations = this.getVariantAnnotations(variants);
-
-        for (VariantAnnotation variantAnnotation : variantAnnotations)
-        {
-            // gets mutation assessor annotation object from variant annotation map
-            MutationAssessor mutationAssessor = this.getMutationAssessorFromEnrichedVariantAnnotation(variantAnnotation);
-
-            if (mutationAssessor != null)
-            {
-                mutationAssessors.add(mutationAssessor);
-            }
-        }
-
-        return mutationAssessors;
-    }
-
-    private MutationAssessor getMutationAssessorFromEnrichedVariantAnnotation(VariantAnnotation variantAnnotation)
-    {
-        // gets mutation assessor annotation object from variant annotation map
-        Map<String, Object> map = variantAnnotation.getDynamicProps();
-
-        MutationAssessorAnnotation mutationAssessorAnnotation
-            = (MutationAssessorAnnotation) map.get("mutation_assessor");
-
-        if (mutationAssessorAnnotation != null)
-        {
-            MutationAssessor obj = mutationAssessorAnnotation.getAnnotation();
-
-            if (obj != null &&
-                obj.getMappingIssue().length() == 0)
-            {
-                return obj;
-            }
-        }
-
-        return null;
-    }
-
-    private EnrichmentService initPostEnrichmentService()
-    {
-        // The post enrichment service enriches the annotation after saving
-        // the original annotation data to the repository. Any enrichment
-        // performed by the post enrichment service is not saved
-        // to the annotation repository.
-        EnrichmentService postEnrichmentService = new VEPEnrichmentService();
-
-        AnnotationEnricher enricher = new MutationAssessorAnnotationEnricher(mutationAssessorService);
-        postEnrichmentService.registerEnricher("mutation_assessor", enricher);
-
-        return postEnrichmentService;
-    }
-
-    private List<VariantAnnotation> getVariantAnnotations(List<String> variants)
-    {
-        EnrichmentService postEnrichmentService = this.initPostEnrichmentService();
-
-        return this.variantAnnotator.getVariantAnnotations(variants, postEnrichmentService);
-    }
-
-    private VariantAnnotation getVariantAnnotation(String variant)
-        throws VariantAnnotationNotFoundException, VariantAnnotationWebServiceException
-    {
-        EnrichmentService postEnrichmentService = this.initPostEnrichmentService();
-
-        return this.variantAnnotator.getVariantAnnotation(variant, postEnrichmentService);
+        return this.mutationAssessorService.getMutationAssessorFromEnrichedVariantAnnotation(variants);
     }
 }
